+++ starsim/networks.py
@@ -108,16 +108,23 @@ class Network(Route):
         """ Relative transmission on each network edge """
         return self.edges['beta'] if 'beta' in self.edges else None
 
+    def init_results(self):
+        """ Store network length by default """
+        super().init_results()
+        self.define_results(
+            ss.Result('n_edges', dtype=int, scale=True, label='Number of edges', auto_plot=False)
+        )
+        return
+
     def init_post(self, add_pairs=True):
         super().init_post()
         if add_pairs: self.add_pairs()
         return
 
     def __len__(self):
-        try:
-            return len(self.edges.p1)
-        except:  # pragma: no cover
-            return 0
+        """ The length is the number of edges """
+        try:    return len(self.edges.p1)
+        except: return 0
 
     def __repr__(self, **kwargs):
         """ Convert to a dataframe for printing """
@@ -235,6 +242,11 @@ class Network(Route):
                 raise KeyError(errormsg)
             self.edges[key] = np.concatenate([curr_arr, new_arr])  # Resize to make room, preserving dtype
         self.validate_uids()
+        return
+
+    def update_results(self):
+        """ Store the number of edges in the network """
+        self.results['n_edges'][self.ti] = len(self)
         return
 
     def to_graph(self): # pragma: no cover